+++ web/src/app/admin/api-key/page.tsx
@@ -31,7 +31,7 @@ import { Button } from "@opal/components";
 import CopyIconButton from "@/refresh-components/buttons/CopyIconButton";
 import Text from "@/refresh-components/texts/Text";
 import { SvgEdit, SvgKey, SvgRefreshCw } from "@opal/icons";
-import { useCloudSubscription } from "@/hooks/useCloudSubscription";
+import { useCloudSubscription } from "@/app/ee/hooks/useCloudSubscription";
 import { ADMIN_ROUTE_CONFIG, ADMIN_PATHS } from "@/lib/admin-routes";
 
 const route = ADMIN_ROUTE_CONFIG[ADMIN_PATHS.API_KEYS]!;
