+++ web/src/app/ee/admin/billing/SubscriptionSummary.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { InfoItem } from "./InfoItem";
-import { statusToDisplay, BillingInformation } from "@/lib/billing";
+import { statusToDisplay, BillingInformation } from "@/lib/ee/billing";
 import { formatDateShort } from "@/lib/dateUtils";
 
 interface SubscriptionSummaryProps {
